cinncinati: Warn when the local deploy isn't found in the graph
When going through the update graph, Zincati searches for a node that
matches the current deployement, as the starting point for the update.

If the graph does not contains a matching node, zincati errors silently.
While this should almost never happen, it's useful to print a warning to
explain the situation if we hit that case.